refactor user controller
