+++ server/controllers/user/login.action.js
@@ -1,0 +1,11 @@
+const UserService = require("../../services/user/user.service");
+
+const userService = new UserService();
+
+exports.login = async (req, res) => {
+  const userDTO = req.body;
+
+  const jwt = await userService.login(userDTO);
+
+  res.send({ jwt });
+};
