+++ server/controllers/user/register.action.js
@@ -1,0 +1,11 @@
+const UserService = require("../../services/user/user.service");
+
+const userService = new UserService();
+
+exports.register = async (req, res) => {
+  const userDTO = req.body;
+
+  const user = await userService.register(userDTO);
+
+  res.send(user);
+};
