+++ server/middleware/validators/user/index.js
@@ -1,0 +1,8 @@
+const { register } = require("./register.validator");
+const { login } = require("./login.validator");
+
+const userValidators = {
+  register,
+  login,
+};
+module.exports = userValidators;
