+++ server/routes/index.js
@@ -1,6 +1,6 @@
 module.exports = function (app) {
   app.use("/api/v1/posts", require("./v1/posts"));
-  app.use("/api/v1/users", require("./v1/user"));
+  app.use("/api/v1/users", require("../controllers/user/user.routing"));
   app.use("/api/v1/comments", require("./v1/comments"));
   app.use("/api/v1/images", require("./v1/images"));
 };
