--- server/routes/v1/user.js
@@ -1,12 +1,0 @@
-const express = require("express");
-const router = express();
-const User = require("../../controllers/user.controller");
-const { authorizeUser } = require("../../middleware/auth.middleware");
-const asyncWrapper = require("../../middleware/async-wrapper.middleware");
-
-router
-  .post("/register", asyncWrapper(User.register))
-  .post("/login", asyncWrapper(User.login))
-  .get("/profile", authorizeUser, asyncWrapper(User.profile));
-
-module.exports = router;
